marioppo/alveybj/requirements.txt
setuptools==65.5.0
gym
gym-notices
gym-retro
matplotlib
numpy
pandas
Pillow
stable-baselines3
sb3-contrib
tensorboard
tqdm
zipp